lazy load web rich text editor
